#23 - restructures package for progress tracker
