+++ src/pages/ProgressTrackersPage/ProgressTracker/ProgressTracker.test.tsx
@@ -1,10 +1,10 @@
 import { render, screen } from '@testing-library/react';
 import ProgressTracker from './ProgressTracker';
-import { Niyam } from '../../config/niyams';
-import useNiyamProgressInfo from '../../hooks/useNiyamProgressInfo';
-import { NiyamBuilder } from '../../../test/testUtils';
+import { Niyam } from '../../../config/niyams';
+import useNiyamProgressInfo from '../../../hooks/useNiyamProgressInfo';
+import { NiyamBuilder } from '../../../../test/testUtils';
 
-jest.mock('../../hooks/useNiyamProgressInfo');
+jest.mock('../../../hooks/useNiyamProgressInfo');
 
 describe('ProgressTracker', () => {
   function renderComponent(niyam: Niyam = Niyam.OradaNaPads) {
