+++ src/pages/ProgressTrackersPage/ProgressTracker/ProgressTracker.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import { Niyam } from '../../config/niyams';
-import { slugify } from '../../utils/string';
 import { Box, Card, Grid, LinearProgress, Typography } from '@mui/material';
-import useNiyamProgressInfo from '../../hooks/useNiyamProgressInfo';
+import { Niyam } from '../../../config/niyams';
+import { slugify } from '../../../utils/string';
+import useNiyamProgressInfo from '../../../hooks/useNiyamProgressInfo';
 
 interface ProgressTrackerProps {
   niyam: Niyam;
@@ -13,7 +13,6 @@ function ProgressTracker(props: ProgressTrackerProps): JSX.Element {
 
   function progressBarValue(progress: number, target: number) {
     const percentage = (progress / target) * 100;
-
     return percentage > 100 ? 100 : percentage;
   }
